requirements.txt
fastapi
uvicorn[standard]
pydantic
python-jose
passlib[bcrypt]
sqlalchemy
databases[asyncpg]
python-dotenv
alembic
pytest
httpx
email-validator